+++ src/advanced/__tests__/advanced.test.tsx
@@ -3,6 +3,7 @@ import { describe, expect, test } from 'vitest';
 import { act, fireEvent, render, screen, within } from '@testing-library/react';
 import { CartPage } from '../../refactoring/pages/CartPage';
 import { AdminPage } from '../../refactoring/pages/AdminPage';
+import { ProductsProvider, CouponsProvider, CartProvider } from '../../refactoring/contexts';
 import { Coupon, Product } from '../../types';
 
 const mockProducts: Product[] = [
@@ -43,39 +44,32 @@ const mockCoupons: Coupon[] = [
   },
 ];
 
-const TestAdminPage = () => {
-  const [products, setProducts] = useState<Product[]>(mockProducts);
-  const [coupons, setCoupons] = useState<Coupon[]>(mockCoupons);
-
-  const handleProductUpdate = (updatedProduct: Product) => {
-    setProducts((prevProducts) =>
-      prevProducts.map((p) => (p.id === updatedProduct.id ? updatedProduct : p)),
-    );
-  };
-
-  const handleProductAdd = (newProduct: Product) => {
-    setProducts((prevProducts) => [...prevProducts, newProduct]);
-  };
-
-  const handleCouponAdd = (newCoupon: Coupon) => {
-    setCoupons((prevCoupons) => [...prevCoupons, newCoupon]);
-  };
-
+const TestCartPage = () => {
   return (
-    <AdminPage
-      products={products}
-      coupons={coupons}
-      onProductUpdate={handleProductUpdate}
-      onProductAdd={handleProductAdd}
-      onCouponAdd={handleCouponAdd}
-    />
+    <ProductsProvider initialProducts={mockProducts}>
+      <CouponsProvider initialCoupons={mockCoupons}>
+        <CartProvider>
+          <CartPage />
+        </CartProvider>
+      </CouponsProvider>
+    </ProductsProvider>
+  );
+};
+
+const TestAdminPage = () => {
+  return (
+    <ProductsProvider initialProducts={mockProducts}>
+      <CouponsProvider initialCoupons={mockCoupons}>
+        <AdminPage />
+      </CouponsProvider>
+    </ProductsProvider>
   );
 };
 
 describe('advanced > ', () => {
   describe('시나리오 테스트 > ', () => {
     test('장바구니 페이지 테스트 > ', async () => {
-      render(<CartPage products={mockProducts} coupons={mockCoupons} />);
+      render(<TestCartPage />);
       const product1 = screen.getByTestId('product-p1');
       const product2 = screen.getByTestId('product-p2');
       const product3 = screen.getByTestId('product-p3');
