+++ src/refactoring/App.tsx
@@ -1,9 +1,9 @@
-import { useState } from 'react';
-import { CartPage } from './pages/CartPage.tsx';
+import { CartPage } from './pages/CartPage';
 import { AdminPage } from './pages/AdminPage.tsx';
 import Header from './components/Header.tsx';
 import { Coupon, Product } from '../types.ts';
-import { useCoupons, useProducts } from './hooks';
+import { useUserRole } from './hooks';
+import { ProductsProvider, CouponsProvider, CartProvider } from './contexts';
 
 const initialProducts: Product[] = [
   {
@@ -48,30 +48,24 @@ const initialCoupons: Coupon[] = [
 ];
 
 const App = () => {
-  const { products, updateProduct, addProduct } = useProducts(initialProducts);
-  const { coupons, addCoupon } = useCoupons(initialCoupons);
-  const [isAdmin, setIsAdmin] = useState(false);
-
-  const handleTogglePage = () => {
-    setIsAdmin(!isAdmin);
-  };
+  const { isAdmin, toggleRole } = useUserRole();
 
   return (
     <div className="min-h-screen bg-gray-100">
-      <Header isAdmin={isAdmin} onTogglePage={handleTogglePage} />
-      <main className="container mx-auto mt-6">
-        {isAdmin ? (
-          <AdminPage
-            products={products}
-            coupons={coupons}
-            onProductUpdate={updateProduct}
-            onProductAdd={addProduct}
-            onCouponAdd={addCoupon}
-          />
-        ) : (
-          <CartPage products={products} coupons={coupons} />
-        )}
-      </main>
+      <Header isAdmin={isAdmin} toggleRole={toggleRole} />
+      <ProductsProvider initialProducts={initialProducts}>
+        <CouponsProvider initialCoupons={initialCoupons}>
+          <main className="container mx-auto mt-6">
+            {isAdmin ? (
+              <AdminPage />
+            ) : (
+              <CartProvider>
+                <CartPage />
+              </CartProvider>
+            )}
+          </main>
+        </CouponsProvider>
+      </ProductsProvider>
     </div>
   );
 };
