+++ src/refactoring/components/layouts/index.ts
@@ -1,0 +1,2 @@
+export { default as PageLayout } from './PageLayout';
+export { default as SectionLayout } from './SectionLayout';
