+++ src/refactoring/components/sections/index.ts
@@ -1,0 +1,4 @@
+export { default as CartSection } from './CartSection';
+export { default as ProductListSection } from './ProductListSection';
+export { default as ProductAdminSection } from './ProductAdminSection';
+export { default as CouponAdminSection } from './CouponAdminSection';
