+++ src/refactoring/contexts/index.tsx
@@ -1,0 +1,3 @@
+export * from './CartContext';
+export * from './CouponsContext';
+export * from './ProductsContext';
