+++ src/refactoring/hooks/useUserRole.ts
@@ -1,0 +1,14 @@
+import { useState, useCallback } from 'react';
+
+export const useUserRole = () => {
+  const [isAdmin, setIsAdmin] = useState(false);
+
+  const toggleRole = useCallback(() => {
+    setIsAdmin((prev) => !prev);
+  }, []);
+
+  return {
+    isAdmin,
+    toggleRole,
+  };
+};
