+++ src/refactoring/utils/product.ts
@@ -1,0 +1,10 @@
+import { Product, CartItem } from '../../types';
+
+export const getMaxDiscount = (product: Product) => {
+  return product.discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+};
+
+export const getRemainingStock = (product: Product, cart: CartItem[]) => {
+  const cartItem = cart.find((item) => item.product.id === product.id);
+  return product.stock - (cartItem?.quantity || 0);
+};
